requirements.txt
tokenizers==0.21.0
transformers==4.47.0
chromadb==0.5.23
sentence-transformers
tqdm
python-dotenv
datasets